Feature: adds plugin loading to mujoco_ros
Plugins must derive MujocoSim::MujocoPlugin and can be loaded via pluginlib. To
request loading a plugin, an entry in the parameter server has to be made
specifying the plugin type to load and optionally provide a rosparam path for
it, i.e. the plugin namespace. For an example see the 'demo.launch' launchfile.