Fixes applying initial joint positions
If initial joint positions were supplied mujoco would crash, because they
were tried to set before the model was loaded. This now moved to an extra
function which is called on model loading and when 'reset' is triggered.
This also allows changing initial joint positions at runtime, which get applied
once a 'reload' or 'reset' is triggered.