Changes how mujoco plugin config path on rosparam server is resolved
The plugin configuration can now also be found, if mujoco_ros is started in a
namespace. 'MujocoPlugins' is searched for upwards the namespace tree, so that
either a single config for each namespace can be defined in the respective
namespace, or a more general one in (one of) the parent namespace(s).