Moves off-screen rendering to simulate thread/env step loop threads.
For this to be possible, the invisible glfw windows have to be created in the
same thread they will be used to render images. This is why the initialization
of all vis struct resources moved to an extra function to be called from said
thread.
In single env mode, initialization is requested via a flag in the settings. In
parallel mode each env step loop thread calls the initialization function in the
beginning.

These changes have the advantage that independent environments can be rendered
in parallel, and the images are rendered in the same thread performing the
physics update, thus ensuring the images are always published at the correct
time without having to synchronize several threads. This makes locking
render_mtx for off-screen rendering and the buffer state enum per env superfluous.